Remove the explicit size argument
BeaconDataExtract provides the size of the extracted data through the
optional size parameter. Using that allows the parameter to be removed
and also eliminates the possibility of a descrepancy where size !=
len(data).

This doesn't make much of a difference for Cobalt Strike users because
the aggressor script handled the value automatically. For Metasploit
users however, it removes the necessity to know the size in bytes of the
.NET executable file. This makes it easier to use since the BOF
arguments have to be explicitly packed.

See:
  https://hstechdocs.helpsystems.com/manuals/cobaltstrike/current/userguide/content/topics/beacon-object-files_main.htm#BeaconDataExtract